adjust get all and create conversation endpoints
